Dev -> Main Merge For v0.1 (#10)
* First base version of login/register system (#1)

Simple routing between pages has been delivered.

After much consideration and reading the Racket documentation and GPT's views on this, I concluded that user objects should be stored as structs rather than objects. This is easier to serialize and deserialize into a SQLite database. The Racket database library is designed to pass data as structures without the need for object instantiation, saving a lot of work and file space.

I have prepared a design of the system structure in the file structure_sketch.txt so that you can see what idea I was guided by.

We don't have to stick to this in the long run, the structure is subject to future modifications, but I needed to have something to start working with.

I divided the project into segments, just as I would in Java. Database contains the file for managing the database service. Models contains objects stored in the database. Controllers contains controllers that navigate the user to individual subpages and their logic.

If you are not familiar with any part of this code, I recommend you use Github Copilot to analyze the project and return a brief summary.

* Added Fan and Creator dashboard (#4)

After logging on as either creator/fan you're redirected to the corresponding dashboard.

* - Added Page Render (#6)

- Applied Render to Home, Login and Register

* Added the concert browse page and the ability to bypass login and browse as a guest. (#7)

* Major update of concert listings system (#8)

* Creator dashboard modifications (#9)

* Added Fan and Creator dashboard

After logging on as either creator/fan you're redirected to the corresponding dashboard.

* Added buttons to creator dashboard + created creator settings page

---------

Co-authored-by: Kamil Gzyl <101363877+cammaj@users.noreply.github.com>
Co-authored-by: GKB02 <151150815+GKB02@users.noreply.github.com>